requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.0.2
opencv-contrib-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
